Separated ResearchObject and WritableBagFile into 2 files
