+++ tests/test_provenance.py
@@ -16,11 +16,13 @@ from rdflib.term import Literal
 
 import cwltool.cwlprov as provenance
 from cwltool.cwlprov import provenance_constants
+from cwltool.cwlprov.writablebagfile import close_ro, write_bag_file # change this later
 from cwltool.main import main
 from cwltool.cwlprov.ro import ResearchObject
 from cwltool.stdfsaccess import StdFsAccess
 
 from .util import get_data, needs_docker, working_directory
+
 
 # RDF namespaces we'll query for later
 ORE = Namespace("http://www.openarchives.org/ore/terms/")
@@ -596,21 +598,22 @@ def check_prov(
 def research_object(tmp_path: Path) -> Generator[ResearchObject, None, None]:
     re_ob = ResearchObject(StdFsAccess(str(tmp_path / "ro")), temp_prefix_ro=str(tmp_path / "tmp"))
     yield re_ob
-    re_ob.close()
+    # re_ob.close()
+    close_ro(re_ob)
 
 
 def test_absolute_path_fails(research_object: ResearchObject) -> None:
     with pytest.raises(ValueError):
-        research_object.write_bag_file("/absolute/path/fails")
+        write_bag_file(research_object, "/absolute/path/fails")
 
 
 def test_climboutfails(research_object: ResearchObject) -> None:
     with pytest.raises(ValueError):
-        research_object.write_bag_file("../../outside-ro")
+        write_bag_file(research_object, "../../outside-ro")
 
 
 def test_writable_string(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("file.txt") as fh:
+    with write_bag_file(research_object, "file.txt") as fh:
         assert fh.writable()
         fh.write("Hello\n")
 
@@ -642,19 +645,19 @@ def test_writable_string(research_object: ResearchObject) -> None:
 
 
 def test_writable_unicode_string(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("file.txt") as fh:
+    with write_bag_file(research_object, "file.txt") as fh:
         assert fh.writable()
         fh.write("Here is a snowman: \u2603 \n")
 
 
 def test_writable_bytes(research_object: ResearchObject) -> None:
     string = "Here is a snowman: \u2603 \n".encode()
-    with research_object.write_bag_file("file.txt", encoding=None) as fh:
+    with write_bag_file(research_object, "file.txt", encoding=None) as fh:
         fh.write(string)  # type: ignore
 
 
 def test_data(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("data/file.txt") as fh:
+    with write_bag_file(research_object, "data/file.txt") as fh:
         assert fh.writable()
         fh.write("Hello\n")
 
@@ -668,21 +671,21 @@ def test_data(research_object: ResearchObject) -> None:
 
 
 def test_not_seekable(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("file.txt") as fh:
+    with write_bag_file(research_object, "file.txt") as fh:
         assert not fh.seekable()
         with pytest.raises(OSError):
             fh.seek(0)
 
 
 def test_not_readable(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("file.txt") as fh:
+    with write_bag_file(research_object, "file.txt") as fh:
         assert not fh.readable()
         with pytest.raises(OSError):
             fh.read()
 
 
 def test_truncate_fails(research_object: ResearchObject) -> None:
-    with research_object.write_bag_file("file.txt") as fh:
+    with write_bag_file(research_object, "file.txt") as fh:
         fh.write("Hello there")
         fh.truncate()  # OK as we're always at end
         # Will fail because the checksum can't rewind
